requirements.txt
torch
pillow
git+https://github.com/openai/CLIP.git